refactor: split domain parser responsibilities
